Initial refactoring to permit the distribution of large Bspline orbital sets to be distributed among multiple MPI ranks.  The datasets are distributed along the orbital access, which each rank holding a contiguous range of orbitals.  To minimize communication bandwidth, orbital evaluation is performed in a cooperative fashion in which positions for evaluation are first shared within an orbital group via an AllGather operation.  Local subsets of orbital values (gradients, Laplacians, and Hessians) are then computed for all positions.  These values are then communicated point-wise back to the rank which originated each position.
This cooperative evaluation requires lock-step coordination between the MPI ranks in an orbital group.  Some modification of the outer driver loops is thus required to avoid deadlocks.  In particular, after the completion of a loop over all electrons in a group a call is make to Psi.completeDistributedEvaluations(), which accounts for any load imbalance.  The loop ordering when computing wave function ratios for NonlocalECPotential calculations has also been reversed so that the loop over electrons is outermost.  This is to avoid a deadlock due to the fact that each spin group has different DistributedBsplineSet instance.

This first commit has the significant limitation that multiple OpenMP threads are not yet supported.  This will be addressed in a subsequent commit.  In addition, hybrid orbital representation (i.e. 3D B-spline + spherical atomic orbitals) are not yet supported.  The AtomicOrbital classes currently make use of distance tables, which complicates remote evaluation.